requirements.txt
# --- Core ML & Diffusers ---
# For CUDA 12.1 and higher
--extra-index-url https://download.pytorch.org/whl/cu121
torch
torchvision
torchaudio
diffusers
transformers
accelerate

# --- Custom Model Backend ---
nunchaku @ git+https://github.com/mit-han-lab/nunchaku.git

# --- Web UI Framework and Server (with pinned versions for stability) ---
gradio
fastapi==0.110.0
uvicorn[standard]==0.29.0

# --- API Clients & Web ---
requests
google-generativeai
python-dotenv
openai
grok

# --- Utilities ---
PyYAML
cryptography
Pillow
numpy
